chore: 검색 결과 query -> slug 변경
