--- apps/client/src/app/search/_components/RecentSearchWord/index.ts
@@ -1,1 +1,0 @@
-export * from "./RecentSearchWord";
